+++ bevy_rapier3d/examples/joints3.rs
@@ -8,11 +8,12 @@ fn main() {
             0xF9 as f32 / 255.0,
             0xFF as f32 / 255.0,
         )))
-        .add_plugins(DefaultPlugins)
-        .add_plugin(RapierPhysicsPlugin::<NoUserData>::default())
-        .add_plugin(RapierDebugRenderPlugin::default())
-        .add_startup_system(setup_graphics)
-        .add_startup_system(setup_physics)
+        .add_plugins((
+            DefaultPlugins,
+            RapierPhysicsPlugin::<NoUserData>::default(),
+            RapierDebugRenderPlugin::default(),
+        ))
+        .add_systems(Startup, (setup_graphics, setup_physics))
         .run();
 }
 
@@ -49,6 +50,37 @@ fn create_prismatic_joints(commands: &mut Commands, origin: Vect, num: usize) {
             .local_anchor2(Vec3::new(0.0, 0.0, -shift))
             .limits([-2.0, 2.0]);
         let joint = ImpulseJoint::new(curr_parent, prism);
+
+        curr_parent = commands
+            .spawn((
+                TransformBundle::from(Transform::from_xyz(origin.x, origin.y, origin.z + dz)),
+                RigidBody::Dynamic,
+                Collider::cuboid(rad, rad, rad),
+                joint,
+            ))
+            .id();
+    }
+}
+
+fn create_rope_joints(commands: &mut Commands, origin: Vect, num: usize) {
+    let rad = 0.4;
+    let shift = 1.0;
+
+    let mut curr_parent = commands
+        .spawn((
+            TransformBundle::from(Transform::from_xyz(origin.x, origin.y, origin.z)),
+            RigidBody::Fixed,
+            Collider::cuboid(rad, rad, rad),
+        ))
+        .id();
+
+    for i in 0..num {
+        let dz = (i + 1) as f32 * shift;
+
+        let rope = RopeJointBuilder::new()
+            .local_anchor2(Vec3::new(0.0, 0.0, -shift))
+            .limits([0.0, 2.0]);
+        let joint = ImpulseJoint::new(curr_parent, rope);
 
         curr_parent = commands
             .spawn((
@@ -243,5 +275,6 @@ pub fn setup_physics(mut commands: Commands) {
     create_prismatic_joints(&mut commands, Vec3::new(20.0, 10.0, 0.0), 5);
     create_revolute_joints(&mut commands, Vec3::new(20.0, 0.0, 0.0), 3);
     create_fixed_joints(&mut commands, Vec3::new(0.0, 10.0, 0.0), 5);
+    create_rope_joints(&mut commands, Vec3::new(30.0, 10.0, 0.0), 5);
     create_ball_joints(&mut commands, 15);
 }