+++ src/geometry/shape_views/heightfield.rs
@@ -3,6 +3,7 @@ use rapier::parry::shape::HeightField;
 pub use rapier::parry::shape::HeightFieldCellStatus;
 
 /// Read-only access to the properties of a heightfield.
+#[derive(Copy, Clone)]
 pub struct HeightFieldView<'a> {
     /// The raw shape from Rapier.
     pub raw: &'a HeightField,
@@ -38,8 +39,8 @@ macro_rules! impl_ref_methods(
             }
 
             /// Index of the cell a point is on after vertical projection.
-            pub fn cell_at_point(&self, pt: Vect) -> Option<usize> {
-                self.raw.cell_at_point(&pt.into())
+            pub fn cell_at_point(&self, point: Vect) -> Option<usize> {
+                self.raw.cell_at_point(&point.into())
             }
 
             /// Iterator through all the segments of this heightfield.
@@ -103,8 +104,8 @@ macro_rules! impl_ref_methods(
             }
 
             /// Index of the cell a point is on after vertical projection.
-            pub fn cell_at_point(&self, pt: Vect) -> Option<(usize, usize)> {
-                self.raw.cell_at_point(&pt.into())
+            pub fn cell_at_point(&self, point: Vect) -> Option<(usize, usize)> {
+                self.raw.cell_at_point(&point.into())
             }
 
             /// Iterator through all the triangles of this heightfield.