+++ src/plugin/systems.rs
@@ -11,13 +11,12 @@ use crate::geometry::{
     ColliderMassProperties, ColliderScale, CollisionGroups, ContactForceEventThreshold, Friction,
     RapierColliderHandle, Restitution, Sensor, SolverGroups,
 };
-use crate::pipeline::CollisionEvent;
+use crate::pipeline::{CollisionEvent, ContactForceEvent};
 use crate::plugin::configuration::{SimulationToRenderTime, TimestepMode};
 use crate::plugin::{RapierConfiguration, RapierContext};
 use crate::prelude::{
-    BevyPhysicsHooks, BevyPhysicsHooksAdapter, CollidingEntities, ContactForceEvent,
-    KinematicCharacterController, KinematicCharacterControllerOutput, PhysicsWorld, Real,
-    RigidBodyDisabled,
+    BevyPhysicsHooks, BevyPhysicsHooksAdapter, CollidingEntities, KinematicCharacterController,
+    KinematicCharacterControllerOutput, MassModifiedEvent, PhysicsWorld, Real, RigidBodyDisabled,
 };
 use crate::utils;
 use bevy::ecs::system::{StaticSystemParam, SystemParamItem};
@@ -138,16 +137,21 @@ pub fn apply_scale(
 
 /// System responsible for applying changes the user made to a collider-related component.
 pub fn apply_collider_user_changes(
-    config: Res<RapierConfiguration>,
     mut context: ResMut<RapierContext>,
-    changed_collider_transforms: Query<
-        (
-            &RapierColliderHandle,
-            &GlobalTransform,
-            Option<&PhysicsWorld>,
-        ),
-        (Without<RapierRigidBodyHandle>, Changed<GlobalTransform>),
-    >,
+    config: Res<RapierConfiguration>,
+    (changed_collider_transforms, parent_query, transform_query): (
+        Query<
+            (
+                Entity,
+                &RapierColliderHandle,
+                &GlobalTransform,
+                Option<&PhysicsWorld>,
+            ),
+            (Without<RapierRigidBodyHandle>, Changed<GlobalTransform>),
+        >,
+        Query<&Parent>,
+        Query<&Transform>,
+    ),
     changed_shapes: Query<
         (&RapierColliderHandle, &Collider, Option<&PhysicsWorld>),
         Changed<Collider>,
@@ -216,30 +220,42 @@ pub fn apply_collider_user_changes(
         ),
         Changed<ColliderMassProperties>,
     >,
-) {
-    for (handle, transform, world_within) in changed_collider_transforms.iter() {
-        let world = get_world(world_within, &mut context);
 
-        if let Some(co) = world.colliders.get_mut(handle.0) {
-            if co.parent().is_none() {
-                co.set_position(utils::transform_to_iso(
-                    &transform.compute_transform(),
-                    world.physics_scale,
-                ))
+    mut mass_modified: EventWriter<MassModifiedEvent>,
+) {
+    for (entity, handle, transform, world_within) in changed_collider_transforms.iter() {
+        let world = get_world(world_within, &mut context);
+        let scale = world.physics_scale;
+
+        if world.collider_parent(entity).is_some() {
+            let (_, collider_position) =
+                collider_offset(entity, &world, &parent_query, &transform_query);
+
+            if let Some(co) = world.colliders.get_mut(handle.0) {
+                co.set_position_wrt_parent(utils::transform_to_iso(&collider_position, scale));
             }
+        } else if let Some(co) = world.colliders.get_mut(handle.0) {
+            co.set_position(utils::transform_to_iso(
+                &transform.compute_transform(),
+                scale,
+            ))
         }
     }
 
     for (handle, shape, world_within) in changed_shapes.iter() {
         let world = get_world(world_within, &mut context);
+        let scale = world.physics_scale;
 
         if let Some(co) = world.colliders.get_mut(handle.0) {
             let mut scaled_shape = shape.clone();
-            scaled_shape.set_scale(
-                shape.scale / world.physics_scale,
-                config.scaled_shape_subdivision,
-            );
-            co.set_shape(scaled_shape.raw.clone())
+            scaled_shape.set_scale(shape.scale / scale, config.scaled_shape_subdivision);
+            co.set_shape(scaled_shape.raw.clone());
+
+            if let Some(body) = co.parent() {
+                if let Some(body_entity) = world.rigid_body_entity(body) {
+                    mass_modified.send(body_entity.into());
+                }
+            }
         }
     }
 
@@ -334,6 +350,12 @@ pub fn apply_collider_user_changes(
                 ColliderMassProperties::Mass(mass) => co.set_mass(*mass),
                 ColliderMassProperties::MassProperties(mprops) => {
                     co.set_mass_properties(mprops.into_rapier(world.physics_scale))
+                }
+            }
+
+            if let Some(body) = co.parent() {
+                if let Some(body_entity) = world.rigid_body_entity(body) {
+                    mass_modified.send(body_entity.into());
                 }
             }
         }
@@ -463,6 +485,7 @@ pub fn apply_rigid_body_user_changes(
     >,
     changed_additional_mass_props: Query<
         (
+            Entity,
             &RapierRigidBodyHandle,
             &AdditionalMassProperties,
             Option<&PhysicsWorld>,
@@ -514,6 +537,8 @@ pub fn apply_rigid_body_user_changes(
         ),
         Changed<RigidBodyDisabled>,
     >,
+
+    mut mass_modified: EventWriter<MassModifiedEvent>,
 ) {
     // Deal with sleeping first, because other changes may then wake-up the
     // rigid-body again.
@@ -551,7 +576,7 @@ pub fn apply_rigid_body_user_changes(
     // This is needed for detecting if the user actually changed the rigid-body
     // transform, or if it was just the change we made in our `writeback_rigid_bodies`
     // system.
-    let transform_changed =
+    let transform_changed_fn =
         |handle: &RigidBodyHandle,
          transform: &GlobalTransform,
          last_transform_set: &HashMap<RigidBodyHandle, GlobalTransform>| {
@@ -568,21 +593,38 @@ pub fn apply_rigid_body_user_changes(
     {
         let world = get_world(world_within, &mut context);
 
+        // Use an Option<bool> to avoid running the check twice.
+        let mut transform_changed = None;
+
         if let Some(interpolation) = interpolation.as_deref_mut() {
-            // Reset the interpolation so we don’t overwrite
-            // the user’s input.
-            interpolation.start = None;
-            interpolation.end = None;
+            transform_changed = transform_changed.or_else(|| {
+                Some(transform_changed_fn(
+                    &handle.0,
+                    global_transform,
+                    &world.last_body_transform_set,
+                ))
+            });
+
+            if transform_changed == Some(true) {
+                // Reset the interpolation so we don’t overwrite
+                // the user’s input.
+                interpolation.start = None;
+                interpolation.end = None;
+            }
         }
 
         if let Some(rb) = world.bodies.get_mut(handle.0) {
+            transform_changed = transform_changed.or_else(|| {
+                Some(transform_changed_fn(
+                    &handle.0,
+                    global_transform,
+                    &world.last_body_transform_set,
+                ))
+            });
+
             match rb.body_type() {
                 RigidBodyType::KinematicPositionBased => {
-                    if transform_changed(
-                        &handle.0,
-                        global_transform,
-                        &world.last_body_transform_set,
-                    ) {
+                    if transform_changed == Some(true) {
                         rb.set_next_kinematic_position(utils::transform_to_iso(
                             &global_transform.compute_transform(),
                             world.physics_scale,
@@ -593,11 +635,7 @@ pub fn apply_rigid_body_user_changes(
                     }
                 }
                 _ => {
-                    if transform_changed(
-                        &handle.0,
-                        global_transform,
-                        &world.last_body_transform_set,
-                    ) {
+                    if transform_changed == Some(true) {
                         rb.set_position(
                             utils::transform_to_iso(
                                 &global_transform.compute_transform(),
@@ -624,7 +662,7 @@ pub fn apply_rigid_body_user_changes(
         }
     }
 
-    for (handle, mprops, world_within) in changed_additional_mass_props.iter() {
+    for (entity, handle, mprops, world_within) in changed_additional_mass_props.iter() {
         let world = get_world(world_within, &mut context);
 
         if let Some(rb) = world.bodies.get_mut(handle.0) {
@@ -639,6 +677,8 @@ pub fn apply_rigid_body_user_changes(
                     rb.set_additional_mass(*mass, true);
                 }
             }
+
+            mass_modified.send(entity.into());
         }
     }
 
@@ -1328,6 +1368,41 @@ fn tp_and_sync_recurse(
     }
 }
 
+/// System responsible for writing updated mass properties back into the [`ReadMassProperties`] component.
+pub fn writeback_mass_properties(
+    mut context: ResMut<RapierContext>,
+    config: Res<RapierConfiguration>,
+
+    mut mass_props: Query<&mut ReadMassProperties>,
+    mut mass_modified: EventReader<MassModifiedEvent>,
+) {
+    if config.physics_pipeline_active {
+        for (_, world) in context.worlds.iter() {
+            let scale = world.physics_scale;
+
+            for entity in mass_modified.iter() {
+                if let Some(handle) = world.entity2body.get(entity).copied() {
+                    if let Some(rb) = world.bodies.get(handle) {
+                        if let Ok(mut mass_props) = mass_props.get_mut(**entity) {
+                            let new_mass_props = MassProperties::from_rapier(
+                                rb.mass_properties().local_mprops,
+                                scale,
+                            );
+
+                            // NOTE: we write the new value only if there was an
+                            //       actual change, in order to not trigger bevy’s
+                            //       change tracking when the values didn’t change.
+                            if mass_props.get() != &new_mass_props {
+                                mass_props.set(new_mass_props);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
 /// System responsible for advancing the physics simulation, and updating the internal state
 /// for scene queries.
 #[allow(clippy::too_many_arguments)]
@@ -1370,16 +1445,6 @@ pub fn step_simulation<Hooks>(
         }
     }
 }
-
-/// NOTE: This currently does nothing in 2D.
-#[cfg(feature = "async-collider")]
-#[cfg(feature = "dim2")]
-pub fn init_async_colliders() {}
-
-/// NOTE: This does nothing in 3D with headless.
-#[cfg(feature = "headless")]
-#[cfg(feature = "dim3")]
-pub fn init_async_colliders() {}
 
 /// System responsible for creating `Collider` components from `AsyncCollider` components if the
 /// corresponding mesh has become available.
@@ -1443,6 +1508,42 @@ pub fn init_async_scene_colliders(
     }
 }
 
+fn collider_offset(
+    entity: Entity,
+    world: &RapierWorld,
+    parent_query: &Query<&Parent>,
+    transform_query: &Query<&Transform>,
+) -> (Option<RigidBodyHandle>, Transform) {
+    let mut body_entity = entity;
+    let mut body_handle = world.entity2body.get(&body_entity).copied();
+    let mut child_transform = Transform::default();
+    while body_handle.is_none() {
+        if let Ok(parent_entity) = parent_query.get(body_entity) {
+            if let Ok(transform) = transform_query.get(body_entity) {
+                child_transform = *transform * child_transform;
+            }
+            body_entity = parent_entity.get();
+        } else {
+            break;
+        }
+
+        body_handle = world.entity2body.get(&body_entity).copied();
+    }
+
+    if body_handle.is_some() {
+        if let Ok(transform) = transform_query.get(body_entity) {
+            let scale_transform = Transform {
+                scale: transform.scale,
+                ..default()
+            };
+
+            child_transform = scale_transform * child_transform;
+        }
+    }
+
+    (body_handle, child_transform)
+}
+
 /// System responsible for creating new Rapier colliders from the related `bevy_rapier` components.
 pub fn init_colliders(
     mut commands: Commands,
@@ -1457,7 +1558,8 @@ pub fn init_colliders(
         Without<RapierColliderHandle>,
     >,
     mut rigid_body_mprops: Query<&mut ReadMassProperties>,
-    parent_query: Query<(&Parent, Option<&Transform>)>,
+    parent_query: Query<&Parent>,
+    transform_query: Query<&Transform>,
 ) {
     for (
         (
@@ -1536,21 +1638,9 @@ pub fn init_colliders(
             builder = builder.contact_force_event_threshold(threshold.0);
         }
 
-        let mut body_entity = entity;
-        let mut body_handle = world.entity2body.get(&body_entity).copied();
-        let mut child_transform = Transform::default();
-        while body_handle.is_none() {
-            if let Ok((parent_entity, transform)) = parent_query.get(body_entity) {
-                if let Some(transform) = transform {
-                    child_transform = *transform * child_transform;
-                }
-                body_entity = parent_entity.get();
-            } else {
-                break;
-            }
-
-            body_handle = world.entity2body.get(&body_entity).copied();
-        }
+        let body_entity = entity;
+        let (body_handle, child_transform) =
+            collider_offset(entity, world, &parent_query, &transform_query);
 
         builder = builder.user_data(entity.to_bits() as u128);
 
@@ -1564,10 +1654,10 @@ pub fn init_colliders(
                 // Inserting the collider changed the rigid-body’s mass properties.
                 // Read them back from the engine.
                 if let Some(parent_body) = world.bodies.get(body_handle) {
-                    mprops.0 = MassProperties::from_rapier(
+                    mprops.set(MassProperties::from_rapier(
                         parent_body.mass_properties().local_mprops,
                         physics_scale,
-                    );
+                    ));
                 }
             }
             handle
@@ -1800,6 +1890,16 @@ pub fn init_joints(
     }
 }
 
+// fn find_world(context: &mut RapierContext) -> &mut RapierWorld {
+//     for (_, world) in context.worlds.iter_mut() {
+//         if let Some(handle) = item_finder(world) {
+//             return Some((world, handle));
+//         }
+//     }
+
+//     None
+// }
+
 fn find_item_and_world<T>(
     context: &mut RapierContext,
     item_finder: impl Fn(&mut RapierWorld) -> Option<T>,
@@ -1838,6 +1938,8 @@ pub fn sync_removals(
     mut removed_sensors: RemovedComponents<Sensor>,
     mut removed_rigid_body_disabled: RemovedComponents<RigidBodyDisabled>,
     mut removed_colliders_disabled: RemovedComponents<ColliderDisabled>,
+
+    mut mass_modified: EventWriter<MassModifiedEvent>,
 ) {
     /*
      * Rigid-bodies removal detection.
@@ -1882,6 +1984,10 @@ pub fn sync_removals(
         if let Some((world, handle)) =
             find_item_and_world(&mut context, |world| world.entity2collider.remove(&entity))
         {
+            if let Some(parent) = world.collider_parent(entity) {
+                mass_modified.send(parent.into());
+            }
+
             world
                 .colliders
                 .remove(handle, &mut world.islands, &mut world.bodies, true);
@@ -1890,6 +1996,10 @@ pub fn sync_removals(
     }
 
     for entity in orphan_colliders.iter() {
+        if let Some(parent) = context.collider_parent(entity) {
+            mass_modified.send(parent.into());
+        }
+
         if let Some((world, handle)) =
             find_item_and_world(&mut context, |world| world.entity2collider.remove(&entity))
         {
@@ -1976,7 +2086,6 @@ pub fn sync_removals(
         }
     }
 
-    // TODO: update mass props after collider removal.
     // TODO: what about removing forces?
 }
 
@@ -2201,7 +2310,7 @@ mod tests {
     fn colliding_entities_updates() {
         let mut app = App::new();
         app.add_event::<CollisionEvent>()
-            .add_system(update_colliding_entities);
+            .add_systems(Update, update_colliding_entities);
 
         let entity1 = app.world.spawn(CollidingEntities::default()).id();
         let entity2 = app.world.spawn(CollidingEntities::default()).id();
@@ -2287,8 +2396,8 @@ mod tests {
     #[cfg(all(feature = "dim3", feature = "async-collider"))]
     fn async_collider_initializes() {
         let mut app = App::new();
-        app.add_plugin(HeadlessRenderPlugin)
-            .add_system(init_async_colliders);
+        app.add_plugins(HeadlessRenderPlugin)
+            .add_systems(Update, init_async_colliders);
 
         let mut meshes = app.world.resource_mut::<Assets<Mesh>>();
         let cube = meshes.add(Cube::default().into());
@@ -2314,8 +2423,10 @@ mod tests {
         use bevy::scene::scene_spawner_system;
 
         let mut app = App::new();
-        app.add_plugin(HeadlessRenderPlugin)
-            .add_system(init_async_scene_colliders.after(scene_spawner_system));
+        app.add_plugins(HeadlessRenderPlugin).add_systems(
+            Update,
+            init_async_scene_colliders.after(scene_spawner_system),
+        );
 
         let mut meshes = app.world.resource_mut::<Assets<Mesh>>();
         let cube_handle = meshes.add(Cube::default().into());
@@ -2359,10 +2470,12 @@ mod tests {
     #[test]
     fn transform_propagation() {
         let mut app = App::new();
-        app.add_plugin(HeadlessRenderPlugin)
-            .add_plugin(TransformPlugin)
-            .add_plugin(TimePlugin)
-            .add_plugin(RapierPhysicsPlugin::<NoUserData>::default());
+        app.add_plugins((
+            HeadlessRenderPlugin,
+            TransformPlugin,
+            TimePlugin,
+            RapierPhysicsPlugin::<NoUserData>::default(),
+        ));
 
         let zero = (Transform::default(), Transform::default());
 
@@ -2418,10 +2531,12 @@ mod tests {
     #[test]
     fn transform_propagation2() {
         let mut app = App::new();
-        app.add_plugin(HeadlessRenderPlugin)
-            .add_plugin(TransformPlugin)
-            .add_plugin(TimePlugin)
-            .add_plugin(RapierPhysicsPlugin::<NoUserData>::default());
+        app.add_plugins((
+            HeadlessRenderPlugin,
+            TransformPlugin,
+            TimePlugin,
+            RapierPhysicsPlugin::<NoUserData>::default(),
+        ));
 
         let zero = (Transform::default(), Transform::default());
 
@@ -2502,16 +2617,18 @@ mod tests {
 
     impl Plugin for HeadlessRenderPlugin {
         fn build(&self, app: &mut App) {
-            app.add_plugin(WindowPlugin::default())
-                .add_plugin(AssetPlugin::default())
-                .add_plugin(ScenePlugin)
-                .add_plugin(RenderPlugin {
+            app.add_plugins((
+                WindowPlugin::default(),
+                AssetPlugin::default(),
+                ScenePlugin::default(),
+                RenderPlugin {
                     wgpu_settings: WgpuSettings {
                         backends: None,
                         ..Default::default()
                     },
-                })
-                .add_plugin(ImagePlugin::default());
+                },
+                ImagePlugin::default(),
+            ));
         }
     }
 }